Drop `ShadowDomSupport` feature.
`ShadowRootBuilder` has default value now.
