+++ src/feature/shadow-dom/index.ts
@@ -1,5 +1,4 @@
 export * from './attach-shadow.decorator';
-export * from './shadow-dom-support.feature';
 export * from './shadow-content-root';
 export * from './shadow-dom-event';
 export * from './shadow-root-builder';
